requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.3
werkzeug>=3.0.0
google-generativeai>=0.5.0
pillow>=10.4.0
fpdf2>=2.7.9
gunicorn>=21.2.0
